+++ examples/text_builder.rb
@@ -25,7 +25,9 @@ text = Clin::Text.new do |t|
 end
 puts text
 
-table = Clin::Text::Table.new outside_border: true do |t|
+table = Clin::Text::Table.new(border: true) do |t|
+  t.align :right, :center, :left
+  # t.column_delimiter ' - ', ' # '
   t.header %w(First Last Email)
 
   t.row %w(Timothee Guerin timothee.guerin@outlook.com)