+++ spec/clin/text/table_spec.rb
@@ -1,0 +1,5 @@
+require 'spec_helper'
+
+RSpec.describe Clin::Text::Table do
+
+end